[refact#factories] Eliminate duplication when creating configurations
Move the dict logic to a typed dataclass (Config) from the `SelectorFactory.create` and `ConstraintFactory.create` methods to a higher level, in `RuleFactory`.

This change avoids the repeated creation of the same configuration objects and makes the signatures of the `create` methods more strict and understandable, since they now directly accept typed `SelectorConfig` and `ConstraintConfig` objects.